+++ libhxcfe/sources/loaders/camputerslynxldf_loader/camputerslynxldf_loader.h
@@ -26,4 +26,4 @@
 */
 
 
-int CAMPUTERSLYNX_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int CAMPUTERSLYNX_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
