+++ libhxcfe/sources/loaders/dpx_loader/dpx_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int DPX_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int DPX_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
