+++ libhxcfe/sources/loaders/hdm_loader/hdm_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int HDM_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int HDM_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
