+++ libhxcfe/sources/loaders/kryofluxstream_loader/kryofluxstream_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int KryoFluxStream_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int KryoFluxStream_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
