+++ libhxcfe/sources/loaders/mfm_loader/mfm_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int MFM_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int MFM_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
