+++ libhxcfe/sources/loaders/oricdsk_loader/oricdsk_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int OricDSK_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int OricDSK_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
