+++ libhxcfe/sources/loaders/sap_loader/sap_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int SAP_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int SAP_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
