+++ libhxcfe/sources/loaders/stx_loader/stx_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int STX_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int STX_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
