+++ libhxcfe/sources/loaders/vegasdsk_loader/vegasdsk_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int VEGASDSK_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int VEGASDSK_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
