+++ libhxcfe/sources/loaders/w30_loader/w30_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int W30_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int W30_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
