+++ libhxcfe/sources/loaders/xml_loader/xml_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int XML_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int XML_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
