+++ libhxcfe/sources/thirdpartylibs/adflib/Lib/Win32/nt4_dev.h
@@ -21,6 +21,6 @@
 
 HANDLE NT4OpenDrive(char *strDrive);
 BOOL NT4CloseDrive(HANDLE hDrv);
-BOOL NT4ReadSector(HANDLE hDrv, long iSect, int iSize, void *lpvoidBuf);
-BOOL NT4WriteSector(HANDLE hDrv, long iSect, int iSize, void *lpvoidBuf);
+BOOL NT4ReadSector(HANDLE hDrv, int32_t iSect, int iSize, void *lpvoidBuf);
+BOOL NT4WriteSector(HANDLE hDrv, int32_t iSect, int iSize, void *lpvoidBuf);
 ULONG NT4GetDriveSize(HANDLE hDrv);
