+++ libhxcfe/sources/thirdpartylibs/adflib/Lib/adf_link.h
@@ -29,7 +29,7 @@
 #include"prefix.h"
 
 PREFIX RETCODE adfBlockPtr2EntryName(struct Volume *vol, SECTNUM nSect, SECTNUM lPar, 
-	char **name, long *size);
+	char **name, int32_t *size);
 
 
 #endif /* ADF_LINK_H */
